Crate km

Windows Kernel Mode library.

Modules

basedef
Kernel-Mode Types.

crt
C runtime library.

debug
Debugger support.

device_object
Device Object.

dpc
Deferred Procedure Calls (DPC).

driver_object
Driver Object.

event
Event Objects.

file_object
File Object.

irp
I/O request packets (IRP).

irql
Interrupt Request Level (IRQL).

macros
Macros for Kernel-Mode drivers.

object
Kernel Objects.

pool
Kernel Mode pools.

rtl
NT runtime routines.

status
NT Status codes.

string
Kernel mode string types.

time
NT Time routines.

Macros

KdPrint!
Macro to send a message to the kernel debugger.

KdPrint_u!
Macro to send a message to the kernel debugger for unsafe blocks.

check_unsafe!