Enum km::pool::POOL_TYPE

pub enum POOL_TYPE {
    NonPagedPool,
    PagedPool,
    NonPagedPoolMustSucceed,
    DontUseThisType,
    NonPagedPoolCacheAligned,
    PagedPoolCacheAligned,
    NonPagedPoolCacheAlignedMustS,
    MaxPoolType,
    NonPagedPoolSession,
    PagedPoolSession,
    NonPagedPoolMustSucceedSession,
    DontUseThisTypeSession,
    NonPagedPoolCacheAlignedSession,
    PagedPoolCacheAlignedSession,
    NonPagedPoolCacheAlignedMustSSession,
}

Specifies the type of system memory to allocate.

Variants

NonPagedPool

Nonpageable system memory, can be accessed from any IRQL.

PagedPool

Pageable system memory, can only be allocated and accessed at IRQL < DISPATCH_LEVEL.

NonPagedPoolMustSucceed

DontUseThisType

NonPagedPoolCacheAligned

Nonpaged pool, aligned on processor cache boundaries.

PagedPoolCacheAligned

Paged pool, aligned on processor cache boundaries.

NonPagedPoolCacheAlignedMustS

MaxPoolType

NonPagedPoolSession

PagedPoolSession

NonPagedPoolMustSucceedSession

DontUseThisTypeSession

NonPagedPoolCacheAlignedSession

PagedPoolCacheAlignedSession

NonPagedPoolCacheAlignedMustSSession