Struct km::string::ANSI_STRING

pub struct ANSI_STRING {
    pub Length: u16,
    pub MaximumLength: u16,
    pub Buffer: *const u8,
}

A counted string used for ANSI strings.

Fields

Length

The length in bytes of the string stored in Buffer.

MaximumLength

The length in bytes of Buffer.

Buffer

Pointer to a buffer used to contain a string of characters.