Struct km::string::UNICODE_STRING

pub struct UNICODE_STRING {
    pub Length: u16,
    pub MaximumLength: u16,
    pub Buffer: *const u16,
}

A counted Unicode string.

Fields

Length

The length in bytes of the string stored in Buffer.

MaximumLength

The length in bytes of Buffer.

Buffer

Pointer to a buffer used to contain a string of wide characters.